+++ CHANGELOG.md
@@ -1,8 +1,12 @@
 # Changelog
 
-<!--next-version-placeholder-->
+## v0.1.3 (27/04/2022)
 
-## v0.1.1 (24/04/2022)
+- __repr__ set right for dataframes with pandas version <= 1.4.0
+- tidyframe __init__ now accepts inputs and passes them to pd.DataFrame
+- Along with accepting a pandas dataframe
+
+## v0.1.2 (24/04/2022)
 
 - minor bugfix where `skimpy` is now optional!